fix refacotr error

+++ packages/ui_component/lib/src/design_token.dart
@@ -64,7 +64,7 @@ enum YrTheme {
 
 class DsForm {
   final Color textColor;
-  final int width;
+  final double width;
 
   const DsForm({required this.textColor, required this.width});
 }
